requirements.txt
streamlit
requests
python-dotenv
openpyxl
pandas
